Rethink options and positional args
The initial tack was to follow Python's argparse and let the user
specify both options and positional arguments as "args" for a
parser. This doesn't feel right. It burdens the schema with complex arg
count rules to correctly handle positional arguments.

This approach does not support schema definition for positional args and
instead lets them pass through to the caller, who is then responsible
for handling them.

Similarly, the program usage lines are now completely in the hands of
user, rather than being something that's automatically generated from a
schema. This supports multiple usage examples, which are, as noted in
the README, often clearer than the complex grammar presented from an
argument schema.