Simple validation support for pos arg counts
This is far from a full blown position arg schema (ala options) but it
provides a direct path to check arg count, which is common enough to
be built into the parser.

The default is to accept any number of arguments.

This will allow handlers to be assertive in the arguments they
expect. For example a `pos_args` spec of {1, 2} would correspond to a
handler definition like this:

    handle([Arg1])       -> ...
    handle([Arg1, Arg2]) -> ...